Dev v3

+++ src/backend/verifier_exports.py
@@ -420,9 +420,8 @@ def export_pdf(data, log, regex, document_info, compress_type, ocrise):
                 file = folder_out + '/' + filename
             compress_file(file, compress_type, log, folder_out, filename, document_info['filename'])
 
-        if not ocrise and not compress_type:
-            if os.path.isfile(file):
-                shutil.copy(file, folder_out + '/' + filename)
+        if not ocrise and not compress_type and os.path.isfile(file):
+            shutil.copy(file, folder_out + '/' + filename)
 
         return folder_out + '/' + filename, 200
     else:
@@ -625,7 +624,6 @@ def construct_with_var(data, document_info, regex, separator=None):
     _data = []
     if isinstance(document_info['datas'], str):
         data_tmp = json.loads(document_info['datas'])
-        document_info['datas'] = {}
         document_info['datas'] = data_tmp
 
     for column in data.split('#'):
